+++ lib/edgedb/pool/connection_supervisor.ex
@@ -13,11 +13,7 @@ defmodule EdgeDB.Pool.ConnectionSupervisor do
 
   @spec start_link(Keyword.t()) :: Supervisor.on_start()
   def start_link(opts \\ []) do
-    with {:ok, pid} <- DynamicSupervisor.start_link(__MODULE__, opts) do
-      # always start with a single child at start
-      start_connection(pid, opts)
-      {:ok, pid}
-    end
+    DynamicSupervisor.start_link(__MODULE__, opts)
   end
 
   @spec start_connection(Supervisor.supervisor(), Keyword.t()) ::
@@ -28,11 +24,10 @@ defmodule EdgeDB.Pool.ConnectionSupervisor do
     conn_mod = opts[:conn][:mod]
     conn_opts = opts[:conn][:opts]
 
-    %{active: current_count} = DynamicSupervisor.count_children(supervisor)
-
+    # connections can die quite frequently, so giving them a not unique name may lead to a collision
     DynamicSupervisor.start_child(
       supervisor,
-      conn(owner, queue, current_count + 1, conn_mod, conn_opts)
+      conn(owner, queue, UUID.uuid4(), conn_mod, conn_opts)
     )
   end
 
@@ -44,16 +39,13 @@ defmodule EdgeDB.Pool.ConnectionSupervisor do
   @impl DynamicSupervisor
   def init(opts \\ []) do
     owner = opts[:owner]
-    conn_opts = opts[:conn][:opts]
-
     send(owner, {:set_connections_supervisor, self()})
-
-    sup_opts = [strategy: :one_for_one] ++ Keyword.take(conn_opts, [:max_restarts, :max_seconds])
-    DynamicSupervisor.init(sup_opts)
+    DynamicSupervisor.init(strategy: :one_for_one)
   end
 
   defp conn(owner, queue, id, conn_mod, opts) do
-    child_opts = [id: {conn_mod, owner, id}] ++ Keyword.take(opts, [:shutdown])
+    child_opts =
+      [id: {conn_mod, owner, id}, restart: :temporary] ++ Keyword.take(opts, [:shutdown])
 
     DBConnection.Connection.child_spec(
       conn_mod,