+++ test/edgedb/connection/authentication_test.exs
@@ -1,14 +1,15 @@
 defmodule Tests.EdgeDB.Connection.AuthenticationTest do
   use Tests.Support.EdgeDBCase
 
+  import ExUnit.CaptureLog
+
   describe "trust authentication with valid params" do
     setup do
       %{
         connection_params: [
-          tls_security: :insecure,
-          max_concurrency: 1,
           user: "edgedb_trust",
-          max_restarts: 0,
+          max_concurrency: 1,
+          tls_security: :insecure,
           show_sensitive_data_on_connection_error: true
         ]
       }
@@ -27,7 +28,7 @@ defmodule Tests.EdgeDB.Connection.AuthenticationTest do
         connection_params: [
           user: "edgedb_scram",
           password: "edgedb_scram_password",
-          max_restarts: 0,
+          max_concurrency: 1,
           show_sensitive_data_on_connection_error: true
         ]
       }
@@ -35,20 +36,16 @@ defmodule Tests.EdgeDB.Connection.AuthenticationTest do
 
     test "login successfully", context do
       assert {:ok, client} = EdgeDB.start_link(context.connection_params)
-
       assert 1 = EdgeDB.query_single!(client, "select 1")
     end
   end
 
   describe "SCRAM authentication without password" do
     setup do
-      Process.flag(:trap_exit, true)
-
       %{
         connection_params: [
           user: "edgedb_scram",
-          backoff_type: :stop,
-          max_restarts: 0,
+          max_concurrency: 1,
           show_sensitive_data_on_connection_error: true
         ]
       }
@@ -57,20 +54,20 @@ defmodule Tests.EdgeDB.Connection.AuthenticationTest do
     test "disconnects", context do
       assert {:ok, client} = EdgeDB.start_link(context.connection_params)
 
-      assert_receive {:EXIT, ^client, :killed}, 500
+      assert capture_log(fn ->
+               assert {:error, %DBConnection.ConnectionError{}} = EdgeDB.query(client, "select 1")
+               assert EdgeDB.Pool.concurrency(client) == 0
+             end) =~ "AuthenticationError: authentication failed"
     end
   end
 
   describe "SCRAM authentication with invalid password" do
     setup do
-      Process.flag(:trap_exit, true)
-
       %{
         connection_params: [
           username: "edgedb_scram",
           password: "wrong",
-          backoff_type: :stop,
-          max_restarts: 0,
+          max_concurrency: 1,
           show_sensitive_data_on_connection_error: true
         ]
       }
@@ -79,7 +76,10 @@ defmodule Tests.EdgeDB.Connection.AuthenticationTest do
     test "disconnects", context do
       assert {:ok, client} = EdgeDB.start_link(context.connection_params)
 
-      assert_receive {:EXIT, ^client, :killed}, 500
+      assert capture_log(fn ->
+               assert {:error, %DBConnection.ConnectionError{}} = EdgeDB.query(client, "select 1")
+               assert EdgeDB.Pool.concurrency(client) == 0
+             end) =~ "AuthenticationError: authentication failed"
     end
   end
 end